% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_prepareSFE.R
\name{addAnnotations}
\alias{addAnnotations}
\title{Add annotation information to SFE object}
\usage{
addAnnotations(sfe, annoInfo, annotationDir)
}
\arguments{
\item{sfe}{}

\item{annoInfo}{}

\item{annoDir}{}
}
\description{
Add annotation information to SFE object
}
