% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinates.R
\name{addCoordinates}
\alias{addCoordinates}
\title{Add coordinates to Object}
\usage{
addCoordinates(object, coordinatesFile, scaleFactor = NULL)
}
\arguments{
\item{coordinatesFile}{path to coordinates file exported from Spot Detector}

\item{scaleFactor}{a scaling factor which can be used if the image file has 
been reduced in size after the coordinates were generated. For example if 
the image to be used is 10% the size of the original factor scaleFactor = 10}

\item{Object}{either a Seurat object or SCE}
}
\value{
Object
}
\description{
Adds output of Spot Detector coordinates to 
Seurat object or SCE object created by createSeurat/createSCE. Details about 
how to use Spot Detector can be found:
https://github.com/SpatialTranscriptomicsResearch/st_spot_detector
}
