% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_prepareSFE.R
\name{additionalMD}
\alias{additionalMD}
\title{Add additional metadata to sfe
This function adds additional sample information to the colData slot of an 
exisiting SFE object. The function requires a SFE object and a 
sample information dataframe which must include a column named "sample_id" 
with ids that have been added to the sfe object}
\usage{
additionalMD(sfe, sampleInfo)
}
\arguments{
\item{sfe}{}

\item{sampleInfo}{}
}
\description{
Add additional metadata to sfe
This function adds additional sample information to the colData slot of an 
exisiting SFE object. The function requires a SFE object and a 
sample information dataframe which must include a column named "sample_id" 
with ids that have been added to the sfe object
}
