% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_PlotFunctions.R
\name{addPlotFeatureSFE}
\alias{addPlotFeatureSFE}
\title{Add Plot Feature}
\usage{
addPlotFeatureSFE(sfe, coords_sample, plot_type, plot_feat, sample_id)
}
\arguments{
\item{coords_sample}{}

\item{plot_type}{}

\item{plot_feat}{}

\item{sample_id}{}
}
\description{
Add plot feature to coordinates
}
