% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_helper_functions.R
\name{clustCells}
\alias{clustCells}
\title{Cluster helper function}
\usage{
clustCells(sfe, K = 30, cluster_function = "leiden", dim_red = "PCA")
}
\arguments{
\item{sfe}{}

\item{K}{}

\item{cluster_function}{}

\item{dim_red}{}
}
\description{
Uses the clusterCells function to add cluster results 
directly to sfe object.
}
