% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenX.R
\name{createVisiumSCE}
\alias{createVisiumSCE}
\title{createVisiumSCE}
\usage{
createVisiumSCE(tenXDir = "../outs", resolution = "Low")
}
\arguments{
\item{tenXDir}{The path to Space Ranger outs directory containing spatial 
directory and filtered_feature_bc_matric}

\item{resolution}{Resolution of the tissue image to be used for plotting. 
Can be either "High", or "Low". Default is "Low".}
}
\description{
A function to select to import 10X data into an SCE object
}
\examples{
tenXDir <- file.path(system.file(package = "Spaniel"), "extdata/outs")
sce <- createVisiumSCE(tenXDir, resolution = "Low")


}
