% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_findDomain.R
\name{domainToColData}
\alias{domainToColData}
\title{Assign spots to histological domain}
\usage{
domainToColData(domain_sp, spot_sp, sfe, domain_name, sample_id)
}
\arguments{
\item{domain_sp}{}

\item{spot_sp}{}

\item{sfe}{}
}
\description{
Assign spots to histological domain
}
