% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_findDomain.R
\name{domainToSFE}
\alias{domainToSFE}
\title{FindDomain}
\usage{
domainToSFE(imgFile, domainName, sampleName, sfe, cln = 3, fll = 12)
}
\arguments{
\item{imgFile}{}

\item{domainName}{}

\item{sampleName}{}

\item{sfe}{}

\item{cln}{}

\item{fll}{}
}
\description{
This function converts painted histological domains created in Image 
processing software (eg Amira, or Photoshops) and converts them into 
spatial polygons.
}
\examples{

j <- 1
domain_name <- annotations_df$domain[j]
img_file <- file.path(annotationDir,annotations_df$jpg[j])
sfe <- domainToSFE(img_file, domain_name, "C01", sfe)
}
