% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_findDomain.R
\name{findAllDomains}
\alias{findAllDomains}
\title{Find all domains}
\usage{
findAllDomains(sfe, annoInfo, annotationDir, cln = 3, fll = 12)
}
\arguments{
\item{sfe}{}

\item{annoInfo}{a data frame with three columns. Thei first column must 
contain a sample_id which matches with sample_ids in the sfe object, 
the second column must contain a histological domain name name, 
the third column must contain the image annotation file name.}

\item{cln}{}

\item{fll}{}
}
\description{
This function finds mulitple histological domains. It takes an annotation 
data frame and sfe as input.
}
