% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_PlotFunctions.R
\name{getCoordinatesSFE}
\alias{getCoordinatesSFE}
\title{GetCoordinates from SFE}
\usage{
getCoordinatesSFE(sfe, sample_id, sr_rotated)
}
\arguments{
\item{sfe}{}
}
\description{
GetCoordinates from SFE
}
