% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_findDomain.R
\name{imageToCoords}
\alias{imageToCoords}
\title{Convert image to spatial polygon}
\usage{
imageToCoords(imgFile, cln, fll)
}
\arguments{
\item{imgFile}{}
}
\description{
Convert image to spatial polygon
}
