% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_label.R
\name{labelSFE}
\alias{labelSFE}
\title{Label SFE for Shiny}
\usage{
labelSFE(sfe, grp_cols, qc_cols, exp_cond_cols)
}
\arguments{
\item{sfe}{}

\item{grp_cols}{}

\item{qc_cols}{}

\item{exp_cond_cols}{}
}
\description{
Label SFE for Shiny
}
