% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_wrap_unwrap_SFE_raster.R
\name{newSpatRasterImage}
\alias{newSpatRasterImage}
\title{Create SpatRasterImage from SpatRaster}
\usage{
newSpatRasterImage(x)
}
\arguments{
\item{x}{}
}
\description{
Create SpatRasterImage from SpatRaster
}
