% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseImage.R
\name{parseImage}
\alias{parseImage}
\title{This function parses a HE image to use as the background for plots}
\usage{
parseImage(ImgFile)
}
\arguments{
\item{ImgFile}{Path to the image file}
}
\value{
A rasterized grob
}
\description{
This function parses a HE image to use as the background for plots
}
\examples{
imgFile = file.path(system.file(package = "Spaniel"),
                    "HE_Rep1_resized.jpg")
img = parseImage(imgFile)
}
