% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_prepareSFE.R
\name{prepareSFE}
\alias{prepareSFE}
\title{Add additional metadata and spot positions to data
This is a wrapper function for the additionalMD function and 
s}
\usage{
prepareSFE(sfe, sampleInfo)
}
\arguments{
\item{sfe}{}

\item{sampleInfo}{}
}
\description{
Add additional metadata and spot positions to data
This is a wrapper function for the additionalMD function and 
s
}
