% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_helper_functions.R
\name{preProcessandClusterSFE}
\alias{preProcessandClusterSFE}
\title{Preprocess the data for clustering}
\usage{
preProcessandClusterSFE(
  sfe,
  ncomponents = 50,
  proportion = 0.1,
  autoPC = TRUE,
  K = 30,
  cluster_function = "leiden",
  dim_red = "PCA"
)
}
\arguments{
\item{sfe}{}

\item{K}{}

\item{cluster_function}{}
}
\description{
A wrapper function to normalise
}
