% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readSeurat}
\alias{readSeurat}
\title{Create a Seurat Object From Spatial Transcriptomics Data}
\usage{
readSeurat(Counts, BarcodeFile, ProjectName = projectName,
                        SectionNumber = sectionNo)
}
\arguments{
\item{Counts}{Raw count matrix or data frame where each row represents a
gene and each column represents barcoded location on a spatial
transcriptomics slide. The columns should be named using the
spot barcode (eg "GTCCGATATGATTGCCGC")}

\item{BarcodeFile}{a tab seperated barcode file supplied by Spatial
Trancscriptomics. The file should contains three column:
The first column contains the Spatial Transcriptomics barcode,
the second and third column equate to the x and y location}

\item{ProjectName}{The name of the project which is stored in the Seurat
Object.}

\item{SectionNumber}{The location of the sample on the slide}
}
\value{
A Seurat Object
}
\description{
This function converts a count matrix into a Seurat object. The barcodes for
each spot are added to the metadata of the Seurat object and are used in
plotting the data.
}
\examples{
## Data is taken from DOI: 10.1126/science.aaf2403
exampleCounts <- readRDS(file.path(system.file(package = "Spaniel"),
                            "extdata/counts.rds"))
exampleBarcodes <- file.path(system.file(package = "Spaniel"),
                            "1000L2_barcodes.txt")
SeuratObj <- readSeurat(exampleCounts,
                        exampleBarcodes,
                        ProjectName = "TestProj",
                        SectionNumber = 1
                        )
}
