% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_shinySpaniel.R
\name{runShinySpaniel_SFE}
\alias{runShinySpaniel_SFE}
\title{Run shinySpaniel}
\usage{
runShinySpaniel_SFE(UI = ui, Server = server)
}
\description{
Run shinySpaniel
}
\examples{
runShinySpaniel_SFE()
}
