% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSpots.R
\name{selectSpots}
\alias{selectSpots}
\title{selectSpots}
\usage{
selectSpots(sObj, imgObj)
}
\arguments{
\item{sObj}{Either a Seurat object (version 3) or a SingleCellExperiment 
object containing barcode coordinates in the metadata (Seurat) or
colData (SingleCellExperiment).}

\item{imgObj}{a ggplot grob (see parseImage function)}
}
\description{
A function to select spots to remove from analysis
}
\examples{
## Run the shiny app:
selectSpots(sObj, imgObj)

Click on the spots to remove from downstream analysis. Once all the spots have 
been selected close the shiny app window. A list of spots is 
stored in a text file called points_to_remove.txt in the working directory.

Once this step has been run a filtered Seurat or SCE object can be created using
removeSpots (see removeSpots for more details)

sObjFiltered <- (sObj)
}
