% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_wrap_unwrap_SFE_raster.R
\name{singleImg}
\alias{singleImg}
\title{Extract each image separately}
\usage{
singleImg(sample_id, sfe)
}
\arguments{
\item{sample_id}{}

\item{sfe}{}
}
\description{
Extract each image separately
}
