% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_PlotFunctions.R
\name{spanielPlot_SFE}
\alias{spanielPlot_SFE}
\title{Spaniel Plot SFE}
\usage{
spanielPlot_SFE(
  sfe,
  plot_type = "metadata",
  plot_feat,
  sample_id,
  ptSize = 2,
  ptSizeMin = 0,
  ptSizeMax = 5,
  colLow = "#ff3300",
  colHigh = "#ffff00"
)
}
\arguments{
\item{sfe}{}

\item{plot_type}{}

\item{plot_feat}{}

\item{sample_id}{}
}
\description{
Spaniel Plot SFE
}
