% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_findDomain.R
\name{spotToSFE}
\alias{spotToSFE}
\title{Add spot positions to sfe object}
\usage{
spotToSFE(sfe, sampleName)
}
\arguments{
\item{sfe}{}

\item{sampleName}{}
}
\description{
Add spot positions to sfe object
}
\examples{
spotToSFE("sample03", sfe)

}
