% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialPlot.R
\name{ST_plot}
\alias{ST_plot}
\title{Spatial Transcriptomics Plot}
\usage{
ST_plot(Object, Grob, PlotType = c("NoGenes",
                                            "CountsPerSpot",
                                            "Cluster",
                                            "Gene"),
                Gene = NULL, ClusterRes = NULL, CustomTitle = NULL,
                ScaleData = TRUE, ShowFilter = NULL, pt.size = 2,
                pt.size.min = 0, pt.size.max = 5)
}
\arguments{
\item{Object}{Either a Seurat object (version 3) or a SingleCellExperiment 
object containing barcode coordinates in the metadata (Seurat) or
colData (SingleCellExperiment).}

\item{Grob}{an grob to be used as the backgound image see(parseImage)}

\item{PlotType}{There are 5 types of plots avaiable:
1) NoGenes - This shows the number of genes per spot 
and uses information from "nFeature_RNA" column of 
Seurat object or "total_features_by_counts" from a 
SingleCellExperiment object.
2) CountsPerSpot - This shows the number of counts per
spot. It uses information from "nCount_RNA" column of 
Seurat object or "total_counts" from a 
singleCellExperiment object.
3) Cluster - This plot is designed to show clustering
results stored in the meta.data or colData of an object
4) Gene - This plot shows the expression of a single 
gene. This plot uses scaled/normalised expressin data 
from the scale.data slot of Seurat object or logcounts 
of a SingleCellExperiment object.
5) Other - A generic plot to plot any column from the
meta.data or colData of an object.}

\item{Gene}{Gene to plot}

\item{ClusterRes}{which cluster resolution to plot}

\item{CustomTitle}{Specify plot title (optional)}

\item{ScaleData}{Show scaled data on plot (default is TRUE)}

\item{ShowFilter}{Logical filter showing pass/fail for spots}

\item{pt.size}{Point size used for cluster plot default is 2}

\item{pt.size.min}{Minimum point size used for QC and Gene Expression plots 
default is 0}

\item{pt.size.max}{Maximum point size used for QC and Gene Expression plots
default is 5}
}
\value{
A ggplot spatial transcriptomics plot
}
\description{
This function overlays information from a Seurat object or 
SingleCellExperiment object containing barcodes  onto a H & E image. 
There are 4 plots available showing a) the number of
genes detected per spot, b) the number of reads detected per spot,
c) clustering results, d) the gene expression of a selected gene.
}
\examples{


## Data is taken from DOI: 10.1126/science.aaf2403
SeuratObj <- readRDS(file.path(system.file(package = "Spaniel"),
                        "extdata/SeuratData.rds"))
imgFile <- readRDS(file.path(system.file(package = "Spaniel"),
                        "extdata/image.rds"))

## Counts per spot with a QC filter
minGenes <- 2000
minUMI <- 300000
filter <- SeuratObj$nFeature_RNA > minGenes &
            SeuratObj$nCount_RNA > minUMI
ST_plot(Object = SeuratObj, Grob = imgFile,
        PlotType = "CountsPerSpot",
        ShowFilter = filter)

## Cluster plot
ST_plot(Object = SeuratObj, Grob = imgFile,
        PlotType = "Cluster",
        ClusterRes = "cluster_RNA_snn_res.0.6")

## Gene plot
ST_plot(Object = SeuratObj, Grob = imgFile,
        PlotType = "Gene",
        Gene = "Nrgn")
}
