% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_wrap_unwrap_SFE_raster.R
\name{wrapSFE}
\alias{wrapSFE}
\title{Wrap SpatRaster objects so that they can be reloaded}
\usage{
wrapSFE(sfe)
}
\arguments{
\item{sfe}{}
}
\description{
Wrap SpatRaster objects so that they can be reloaded
}
