% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_helper_functions.R
\name{clustCells}
\alias{clustCells}
\title{Cluster helper function}
\usage{
clustCells(sfe, K = 30, cluster_function = "leiden", dim_red = "PCA")
}
\arguments{
\item{sfe}{a SpatialFeatureExperiment object with log normalised counts}

\item{K}{the number of nearest neighbours to use during graph construction}

\item{cluster_function}{any igraph community detection algorithm}

\item{dim_red}{the reduced dimensions to use for clustering}
}
\description{
Uses the clusterCells function to add cluster results 
directly to sfe object.
}
