% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addClusterCols.R
\name{markClusterCol}
\alias{markClusterCol}
\title{markClusterCol
A function to mark the columns containing cluster information in the metadata
or colData of a Seurat or SCE object. Columns are marked with "cluster_" 
prefix.}
\usage{
markClusterCol(Object, Pattern)
}
\arguments{
\item{Object}{Either a Seurat or SCE object containing clustering information}

\item{Pattern}{pattern indicating which columns contain cluster information}
}
\description{
markClusterCol
A function to mark the columns containing cluster information in the metadata
or colData of a Seurat or SCE object. Columns are marked with "cluster_" 
prefix.
}
\examples{
seuratObj <- markClusterCol(seuratObj, "res")

}
