% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseImage.R
\name{parseImage}
\alias{parseImage}
\title{This function parses a HE image to use as the background for plots}
\usage{
parseImage(imgFile, imgType = "jpg")
}
\arguments{
\item{imgFile}{Path to the image file}

\item{imgType}{Type of image options jpg (default), png}
}
\value{
A rasterized grob
}
\description{
This function parses a HE image to use as the background for plots
}
\examples{
imgFile <- file.path(system.file(package = "Spaniel"),
                    "extdata/outs/spatial/tissue_lowres_image.png")
img <- parseImage(imgFile, imgType = "png")
}
