% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfe_helper_functions.R
\name{preProcess}
\alias{preProcess}
\title{Preprocess the data for clustering}
\usage{
preProcess(
  sfe,
  ncomponents = 50,
  proportion = 0.1,
  autoPC = TRUE,
  dim_red = "PCA"
)
}
\arguments{
\item{sfe}{}

\item{ncomponents}{}

\item{proportion}{}

\item{autoPC}{}
}
\description{
A wrapper function to normalise
}
